requirements.txt
tensorflow-gpu==1.15.0
Cython
contextlib2
pillow
lxml
jupyter
matplotlib
tf_slim
pycocotools
pandas
numpy
requests